[working] making member and mailsend

+++ src/main/java/com/kfq/fund/dao/INoticeDAO.java
@@ -1,6 +1,5 @@
 package com.kfq.fund.dao;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 import com.kfq.fund.vo.NoticeVO;
@@ -10,5 +9,5 @@ public interface INoticeDAO {
 	public List<NoticeVO> getNoticeList(Pagination pagination);
 	public NoticeVO getNoticeByNo(String no);
 	public int getNoticeListCnt();
-	public BigDecimal insertNotice(NoticeVO notice);
+	public int insertNotice(NoticeVO notice);
 }
